# feat(proto): add GZIP pooling