--- a/go.mod
+++ b/go.mod
@@ -11,6 +11,7 @@ require (
 	github.com/gotd/tl v0.4.0
 	github.com/gotd/xor v0.1.1
 	github.com/k0kubun/pp/v3 v3.0.7
+	github.com/klauspost/compress v1.10.3
 	github.com/pkg/errors v0.9.1 // indirect
 	github.com/quasilyte/go-ruleguard/dsl v0.3.3
 	github.com/rogpeppe/go-internal v1.8.0