--- a/internal/proto/gzip.go
+++ b/internal/proto/gzip.go
@@ -4,11 +4,60 @@ import (
 	"bytes"
 	"compress/gzip"
 	"io"
+	"sync"
 
+	"go.uber.org/multierr"
 	"golang.org/x/xerrors"
 
 	"github.com/gotd/td/bin"
 )
+
+type gzipPool struct {
+	writers sync.Pool
+	readers sync.Pool
+}
+
+func newGzipPool() *gzipPool {
+	return &gzipPool{
+		writers: sync.Pool{
+			New: func() interface{} {
+				return gzip.NewWriter(nil)
+			},
+		},
+		readers: sync.Pool{},
+	}
+}
+
+func (g *gzipPool) GetWriter(w io.Writer) *gzip.Writer {
+	writer := g.writers.Get().(*gzip.Writer)
+	writer.Reset(w)
+	return writer
+}
+
+func (g *gzipPool) PutWriter(w *gzip.Writer) {
+	g.writers.Put(w)
+}
+
+func (g *gzipPool) GetReader(r io.Reader) (*gzip.Reader, error) {
+	reader, ok := g.readers.Get().(*gzip.Reader)
+	if !ok {
+		r, err := gzip.NewReader(r)
+		if err != nil {
+			return nil, err
+		}
+		return r, nil
+	}
+
+	if err := reader.Reset(r); err != nil {
+		g.readers.Put(reader)
+		return nil, err
+	}
+	return reader, nil
+}
+
+func (g *gzipPool) PutReader(w *gzip.Reader) {
+	g.readers.Put(w)
+}
 
 // GZIP represents a Packed Object.
 //
@@ -21,14 +70,36 @@ type GZIP struct {
 // GZIPTypeID is TL type id of GZIP.
 const GZIPTypeID = 0x3072cfa1
 
+var (
+	gzipRWPool  = newGzipPool()
+	gzipBufPool = sync.Pool{New: func() interface{} {
+		return bytes.NewBuffer(nil)
+	}}
+	copyBufPool = bin.NewPool(0)
+)
+
 // Encode implements bin.Encoder.
-func (g GZIP) Encode(b *bin.Buffer) error {
+func (g GZIP) Encode(b *bin.Buffer) (rErr error) {
 	b.PutID(GZIPTypeID)
 
 	// Writing compressed data to buf.
-	buf := new(bytes.Buffer)
-	w := gzip.NewWriter(buf)
-	if _, err := io.Copy(w, bytes.NewReader(g.Data)); err != nil {
+	buf := gzipBufPool.Get().(*bytes.Buffer)
+	buf.Reset()
+	defer gzipBufPool.Put(buf)
+
+	// Copy buffer.
+	copyBuf := copyBufPool.GetSize(b.Len())
+	defer copyBufPool.Put(copyBuf)
+
+	w := gzipRWPool.GetWriter(buf)
+	defer func() {
+		if closeErr := w.Close(); closeErr != nil {
+			closeErr = xerrors.Errorf("close: %w", closeErr)
+			multierr.AppendInto(&rErr, closeErr)
+		}
+		gzipRWPool.PutWriter(w)
+	}()
+	if _, err := io.CopyBuffer(w, bytes.NewReader(g.Data), copyBuf.Buf); err != nil {
 		return xerrors.Errorf("compress: %w", err)
 	}
 	if err := w.Close(); err != nil {
@@ -42,7 +113,7 @@ func (g GZIP) Encode(b *bin.Buffer) error {
 }
 
 // Decode implements bin.Decoder.
-func (g *GZIP) Decode(b *bin.Buffer) error {
+func (g *GZIP) Decode(b *bin.Buffer) (rErr error) {
 	if err := b.ConsumeID(GZIPTypeID); err != nil {
 		return err
 	}
@@ -51,11 +122,17 @@ func (g *GZIP) Decode(b *bin.Buffer) error {
 		return err
 	}
 
-	r, err := gzip.NewReader(bytes.NewReader(buf))
+	r, err := gzipRWPool.GetReader(bytes.NewReader(buf))
 	if err != nil {
 		return xerrors.Errorf("gzip error: %w", err)
 	}
-	defer func() { _ = r.Close() }()
+	defer func() {
+		if closeErr := r.Close(); closeErr != nil {
+			closeErr = xerrors.Errorf("close: %w", closeErr)
+			multierr.AppendInto(&rErr, closeErr)
+		}
+		gzipRWPool.PutReader(r)
+	}()
 
 	// Apply mitigation for reading too much data which can result in OOM.
 	const maxUncompressedSize = 1024 * 1024 * 10 // 10 mb