--- a/internal/proto/gzip_test.go
+++ b/internal/proto/gzip_test.go
@@ -2,11 +2,15 @@ package proto
 
 import (
 	"bytes"
+	"crypto/rand"
+	"fmt"
+	"io"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 
 	"github.com/gotd/td/bin"
+	"github.com/gotd/td/internal/testutil"
 )
 
 func TestGZIP_Encode(t *testing.T) {
@@ -34,4 +38,52 @@ func TestGZIP_Decode(t *testing.T) {
 	// TODO(ernado): fail explicitly if limit is reached
 	require.NoError(t, b.Decode(&decoded))
 	require.Less(t, len(decoded.Data), len(g.Data))
+}
+
+func benchmarkGZIPEncode(payloadSize int) func(b *testing.B) {
+	return func(b *testing.B) {
+		g := &GZIP{Data: make([]byte, payloadSize)}
+		_, err := io.ReadFull(rand.Reader, g.Data)
+		require.NoError(b, err)
+
+		var buf bin.Buffer
+		b.ReportAllocs()
+		b.SetBytes(int64(payloadSize))
+		b.ResetTimer()
+
+		for i := 0; i < b.N; i++ {
+			buf.Reset()
+			_ = g.Encode(&buf)
+		}
+	}
+}
+
+func BenchmarkGZIP_Encode(b *testing.B) {
+	for _, size := range testutil.Payloads() {
+		b.Run(fmt.Sprintf("%db", size), benchmarkGZIPEncode(size))
+	}
+}
+
+func benchmarkGZIPDecode(payloadSize int) func(b *testing.B) {
+	return func(b *testing.B) {
+		g := &GZIP{Data: make([]byte, payloadSize)}
+		_, err := io.ReadFull(rand.Reader, g.Data)
+		require.NoError(b, err)
+
+		var buf bin.Buffer
+		require.NoError(b, g.Encode(&buf))
+		b.ReportAllocs()
+		b.SetBytes(int64(payloadSize))
+		b.ResetTimer()
+
+		for i := 0; i < b.N; i++ {
+			_ = g.Decode(&bin.Buffer{Buf: buf.Buf})
+		}
+	}
+}
+
+func BenchmarkGZIP_Decode(b *testing.B) {
+	for _, size := range testutil.Payloads() {
+		b.Run(fmt.Sprintf("%db", size), benchmarkGZIPDecode(size))
+	}
 }